use typescript for vuetify options
